moved search endpoiunt to snowflakes
