+++ src/snovault/__init__.py
@@ -105,7 +105,7 @@ def main(global_config, **local_config):
     if 'elasticsearch.server' in config.registry.settings:
         config.include('snovault.elasticsearch')
         # needed for /search/?
-        config.include('snovault.viewconfigs.views')
+        config.include('snowflakes.viewconfigs.views')
 
     config.include(static_resources)
     config.include(changelogs)
