+++ src/snowflakes/__init__.py
@@ -95,7 +95,6 @@ def main(global_config, **local_config):
 
     if 'elasticsearch.server' in config.registry.settings:
         config.include('snovault.elasticsearch')
-        config.include('snovault.viewconfigs.views')
         config.include('snowflakes.viewconfigs.views')
 
     config.include(static_resources)
